// generated/mysqlx_resultset_pb.ts
export type ColumnMetaData = Message<"Mysqlx.Resultset.ColumnMetaData"> & {
  /**
   * * datatype of the field in a row
   *
   * @generated from field: required Mysqlx.Resultset.ColumnMetaData.FieldType type = 1;
   */
  type: ColumnMetaData_FieldType;

  /**
   * * name of the column
   *
   * @generated from field: optional bytes name = 2;
   */
  name: Uint8Array;

  /**
   * * name of the column before an alias was applied
   *
   * @generated from field: optional bytes original_name = 3;
   */
  originalName: Uint8Array;

  /**
   * * name of the table the column originates from
   *
   * @generated from field: optional bytes table = 4;
   */
  table: Uint8Array;

  /**
   * * name of the table the column originates from before an alias was applied
   *
   * @generated from field: optional bytes original_table = 5;
   */
  originalTable: Uint8Array;

  /**
   * * schema the column originates from
   *
   * @generated from field: optional bytes schema = 6;
   */
  schema: Uint8Array;

  /**
   * * catalog the schema originates from
   * @note
   * As there is currently no support for catalogs in MySQL,
   * don't expect this field to be set. In the MySQL C/S
   * protocol the field had the value ``def`` all the time
   *
   * @generated from field: optional bytes catalog = 7;
   */
  catalog: Uint8Array;

  /**
   * @generated from field: optional uint64 collation = 8;
   */
  collation: bigint;

  /**
   * * displayed factional decimal digits for floating point and
   * fixed point numbers
   *
   * @generated from field: optional uint32 fractional_digits = 9;
   */
  fractionalDigits: number;

  /**
   * * maximum count of displayable characters of .type
   *
   * @generated from field: optional uint32 length = 10;
   */
  length: number;

  /**
   * * ``.type`` specific flags
   * | Type    | Value  | Description  |
   * |---------|--------|--------------|
   * | UINT    | 0x0001 | zerofill     |
   * | DOUBLE  | 0x0001 | unsigned     |
   * | FLOAT   | 0x0001 | unsigned     |
   * | DECIMAL | 0x0001 | unsigned     |
   * | BYTES   | 0x0001 | rightpad     |
   * | Value  | Description     |
   * |--------|-----------------|
   * | 0x0010 | NOT\_NULL       |
   * | 0x0020 | PRIMARY\_KEY    |
   * | 0x0040 | UNIQUE\_KEY     |
   * | 0x0080 | MULTIPLE\_KEY   |
   * | 0x0100 | AUTO\_INCREMENT |
   * default: 0
   *
   * @generated from field: optional uint32 flags = 11;
   */
  flags: number;

  /**
   * * a hint about the higher-level encoding of a BYTES field
   * | Type   | Value  | Description             |
   * |--------|--------|-------------------------|
   * | BYTES  | 0x0001 | GEOMETRY (WKB encoding) |
   * | BYTES  | 0x0002 | JSON (text encoding)    |
   * | BYTES  | 0x0003 | XML (text encoding)     |
   * @note
   * This list isn't comprehensive. As a guideline: the field's
   * value is expected to pass a validator check on client
   * and server if this field is set. If the server adds more
   * internal data types that rely on BLOB storage like image
   * manipulation, seeking into complex types in BLOBs, and
   * more types will be added
   *
   * @generated from field: optional uint32 content_type = 12;
   */
  contentType: number;
};

/**
 * Describes the message Mysqlx.Resultset.ColumnMetaData.
 * Use `create(ColumnMetaDataSchema)` to create a new message.
 */
export const ColumnMetaDataSchema: GenMessage<ColumnMetaData> /*@__PURE__*/ = messageDesc(
  file_mysqlx_resultset,
  4,
);

/**
 * @generated from enum Mysqlx.Resultset.ColumnMetaData.FieldType
 */
export enum ColumnMetaData_FieldType {
  /**
   * @generated from enum value: SINT = 1;
   */
  SINT = 1,

  /**
   * @generated from enum value: UINT = 2;
   */
  UINT = 2,

  /**
   * @generated from enum value: DOUBLE = 5;
   */
  DOUBLE = 5,

  /**
   * @generated from enum value: FLOAT = 6;
   */
  FLOAT = 6,

  /**
   * @generated from enum value: BYTES = 7;
   */
  BYTES = 7,

  /**
   * @generated from enum value: TIME = 10;
   */
  TIME = 10,

  /**
   * @generated from enum value: DATETIME = 12;
   */
  DATETIME = 12,

  /**
   * @generated from enum value: SET = 15;
   */
  SET = 15,

  /**
   * @generated from enum value: ENUM = 16;
   */
  ENUM = 16,

  /**
   * @generated from enum value: BIT = 17;
   */
  BIT = 17,

  /**
   * @generated from enum value: DECIMAL = 18;
   */
  DECIMAL = 18,
}
// ...
export enum ContentType_BYTES {
  /**
   * @generated from enum value: GEOMETRY = 1;
   */
  GEOMETRY = 1,

  /**
   * @generated from enum value: JSON = 2;
   */
  JSON = 2,

  /**
   * @generated from enum value: XML = 3;
   */
  XML = 3,
}